requirements.txt
pytorch
torchvision~=0.11.1
torchaudio
tensorboard
numpy~=1.21.0
matplotlib~=3.4.2
torchviz~=0.0.2
tabulate
dataclasses~=0.8
pillow~=8.4.0
graphviz~=2.38